extract series column
